+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## 0.3.4
 
+* `Core`:
+  * `Common`:
+    * Fixes in hierarchy saver/loader
+    * All `restoreHierarchy` extensions now not suspendable
+  * `JS`:
+    * Experimentally added `OneParameterUrlNavigationConfigsRepo`
+
 ## 0.3.3
 
 * `MVVM`: